+++ app/src/main/java/org/oxycblt/auxio/detail/recycler/AlbumDetailAdapter.kt
@@ -45,7 +45,7 @@ import org.oxycblt.auxio.util.textSafe
 class AlbumDetailAdapter(listener: Listener) :
     DetailAdapter<AlbumDetailAdapter.Listener>(listener, DIFFER) {
     private var currentSong: Song? = null
-    private var currentHighlightedSongPos: Int? = null
+    private var currentSongPos: Int? = null
 
     override fun getCreatorFromItem(item: Item) =
         super.getCreatorFromItem(item)
@@ -65,13 +65,19 @@ class AlbumDetailAdapter(listener: Listener) :
                 else -> null
             }
 
-    override fun onBind(viewHolder: RecyclerView.ViewHolder, item: Item, listener: Listener) {
-        super.onBind(viewHolder, item, listener)
-
-        when (item) {
-            is Album -> (viewHolder as AlbumDetailViewHolder).bind(item, listener)
-            is DiscHeader -> (viewHolder as DiscHeaderViewHolder).bind(item, Unit)
-            is Song -> (viewHolder as AlbumSongViewHolder).bind(item, listener)
+    override fun onBind(
+        viewHolder: RecyclerView.ViewHolder,
+        item: Item,
+        listener: Listener,
+        payload: List<Any>
+    ) {
+        super.onBind(viewHolder, item, listener, payload)
+        if (payload.isEmpty()) {
+            when (item) {
+                is Album -> (viewHolder as AlbumDetailViewHolder).bind(item, listener)
+                is DiscHeader -> (viewHolder as DiscHeaderViewHolder).bind(item, Unit)
+                is Song -> (viewHolder as AlbumSongViewHolder).bind(item, listener)
+            }
         }
     }
 
@@ -83,8 +89,8 @@ class AlbumDetailAdapter(listener: Listener) :
     fun highlightSong(song: Song?) {
         if (song == currentSong) return
         currentSong = song
-        currentHighlightedSongPos?.let { notifyItemChanged(it, PAYLOAD_HIGHLIGHT_CHANGED) }
-        currentHighlightedSongPos = highlightItem(song)
+        currentSongPos?.let { pos -> notifyItemChanged(pos, PAYLOAD_HIGHLIGHT_CHANGED) }
+        currentSongPos = highlightItem(song)
     }
 
     companion object {