+++ app/src/main/java/org/oxycblt/auxio/util/FrameworkUtil.kt
@@ -169,7 +169,11 @@ fun Fragment.launch(
     viewLifecycleOwner.lifecycleScope.launch { viewLifecycleOwner.repeatOnLifecycle(state, block) }
 }
 
-/** Combines the called flow with the given flow and then collects them both into [block]. */
+/**
+ * Combines the called flow with the given flow and then collects them both into [block].
+ * This is a bit of a dumb hack with [combine], as when we have to combine flows, we often
+ * just want to call the same block with both functions, and not do any transformations.
+ */
 suspend fun <T1, T2> Flow<T1>.collectWith(other: Flow<T2>, block: suspend (T1, T2) -> Unit) {
     combine(this, other) { a, b -> a to b }.collect { block(it.first, it.second) }
 }