+++ CHANGELOG.md
@@ -9,9 +9,11 @@
 - Fixed crash when seeking to the end of a track as the track changed to a track with a lower duration 
 - Fixed regression where GadgetBridge media controls would no longer work
 - Fixed issue where the album/artist/genre would not be correctly restored
+- Fixed issue where items would not highlight properly in the detail UI
 
 #### Dev/Meta
 - Switched from `LiveData` to `StateFlow`
+- Use `notifyItemChanged` instead of directly mutating `ViewHolder` instances.
 
 ## v2.3.0